fix(mobile): drop custom header chrome on workspace page for native title + toolbar menu
- plain native centered title (no GlassHeaderTitle pill) on the
  transparent header
- ellipsis becomes a native Stack.Toolbar.Menu (copy branch / rename)
  composed on the (tabs) Stack.Screen, replacing the custom GlassView
  button + ActionSheetIOS
