+++ apps/mobile/app/(authenticated)/workspace/[id]/_layout.tsx
@@ -1,14 +1,12 @@
 import { isLiquidGlassAvailable } from "expo-glass-effect";
 import { Stack, useLocalSearchParams } from "expo-router";
 import { useWorkspaceHost } from "@/hooks/useWorkspaceHost";
-import { GlassHeaderTitle } from "@/screens/(authenticated)/workspace/[id]/chat/components/GlassHeaderTitle";
-import { WorkspaceMenuButton } from "@/screens/(authenticated)/workspace/[id]/components/WorkspaceMenuButton";
+import { WorkspaceToolbarMenu } from "@/screens/(authenticated)/workspace/[id]/components/WorkspaceToolbarMenu";
 
-// Same glass treatment as the chat headers: transparent bar, floating back
-// circle on iOS 26, title carrying its own glass pill.
+// Same glass treatment as the chat headers: transparent bar with the native
+// centered title and a floating back circle on iOS 26.
 const glassHeaderOptions = {
 	headerShown: true,
-	title: "",
 	headerTransparent: true,
 	headerBackButtonDisplayMode: "minimal",
 	headerShadowVisible: false,
@@ -26,16 +24,10 @@ export default function WorkspaceLayout() {
 		<Stack screenOptions={{ headerShown: false }}>
 			<Stack.Screen
 				name="(tabs)"
-				options={{
-					...glassHeaderOptions,
-					headerTitle: () => (
-						<GlassHeaderTitle title={workspace?.name ?? null} />
-					),
-					headerRight: () => (
-						<WorkspaceMenuButton workspace={workspace} host={host} />
-					),
-				}}
-			/>
+				options={{ ...glassHeaderOptions, title: workspace?.name ?? "" }}
+			>
+				<WorkspaceToolbarMenu workspace={workspace} host={host} />
+			</Stack.Screen>
 		</Stack>
 	);
 }
