--- apps/mobile/screens/(authenticated)/workspace/[id]/components/WorkspaceMenuButton/index.ts
@@ -1,1 +1,0 @@
-export { WorkspaceMenuButton } from "./WorkspaceMenuButton";
