+++ apps/mobile/screens/(authenticated)/workspace/[id]/components/WorkspaceToolbarMenu/index.ts
@@ -1,0 +1,1 @@
+export { WorkspaceToolbarMenu } from "./WorkspaceToolbarMenu";
